+++ src/main/java/subway/repository/StationRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.entity.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
