+++ src/main/java/subway/service/StationService.java
@@ -1,7 +1,11 @@
-package subway;
+package subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.repository.StationRepository;
+import subway.dto.StationRequest;
+import subway.dto.StationResponse;
+import subway.entity.Station;
 
 import java.util.List;
 import java.util.stream.Collectors;
